+++ R/r_repos_data.R
@@ -5,6 +5,8 @@
 #' @param include A subset of packages to return data for.
 #' @param cast Cast the results to wide format
 #'  so that each package only appears in one row.
+#' @param add_hex Search for hex sticker URLs for each package.
+#' If the link does not exist, returns \code{NA}.
 #' @param nThread Number of threads to parallelise \code{add_descriptions} 
 #' step across.
 #' @param verbose Print messages.
@@ -12,26 +14,31 @@
 #' @inheritParams github_files
 #' @inheritParams description_extract
 #' @inheritParams BiocManager::repositories
+#' @inheritParams rworkflows::get_hex
 #' @returns data.table
 #' 
 #' @export
 #' @importFrom utils installed.packages available.packages
-#' @importFrom data.table merge.data.table
+#' @importFrom data.table merge.data.table := fcoalesce
+#' @importFrom rworkflows get_hex
 #' @examples 
 #' #### All packages ####
 #' pkgs1 <- r_repos_data()
 #' #### Specific packages ####
 #' #### Add extra data ####
-#' pkgs3 <- r_repos_data(include="echogithub",
+#' pkgs2 <- r_repos_data(include=c("echogithub","echodeps","BiocManager"),
 #'                       which = r_repos_opts(exclude = NULL),
 #'                       add_downloads = TRUE,
 #'                       add_descriptions = TRUE,
 #'                       add_github = TRUE,
-#'                       cast=TRUE)
+#'                       add_hex = TRUE,
+#'                       cast = TRUE)
 r_repos_data <- function(include=NULL,
                          add_downloads=FALSE,
                          add_descriptions=FALSE,
                          add_github=FALSE,
+                         add_hex=FALSE,
+                         hex_path="inst/hex/hex.png",
                          which=r_repos_opts(),
                          cast=FALSE,
                          fields=NULL,
@@ -40,7 +47,7 @@ r_repos_data <- function(include=NULL,
                          token=gh::gh_token(),
                          verbose=TRUE){ 
     # devoptera::args2vars(r_repos_data, reassign = TRUE)
-    installed <- package <- NULL;
+    package <- ref <- hex_url <- NULL;
     
     if(isTRUE(add_github) && 
        isFALSE(add_descriptions)){
@@ -69,17 +76,19 @@ r_repos_data <- function(include=NULL,
     }
     #### Add DESRIPTION metadata ####
     if(isTRUE(add_descriptions)){
-        meta_desc <- description_extract_multi(pkgs = unique(pkgs$package),
-                                               fields = fields,
-                                               nThread = nThread,
-                                               verbose = verbose)
+        meta_desc <- description_extract(refs = unique(pkgs$package),
+                                         fields = fields,
+                                         as_datatable = TRUE,
+                                         nThread = nThread,
+                                         verbose = verbose)
         by <- base::intersect(names(meta_desc),names(pkgs))
         pkgs <- data.table::merge.data.table(meta_desc,
                                              pkgs,
                                              all = TRUE,
                                              by=by)
     }
-    #### Add GitHub metadata ####
+    #### Add owner_repo ####
+    pkgs <- add_owner_repo(dt = pkgs)
     #### Get repo metadata ####
     if(isTRUE(add_github)){
         if(!all(c("owner","repo") %in% names(pkgs))){
@@ -93,7 +102,7 @@ r_repos_data <- function(include=NULL,
                     gm <- github_metadata(
                         owner = pkgs_p$owner,
                         repo = pkgs_p$repo,
-                        add_traffic = TRUE,
+                        add_traffic = FALSE,
                         token = token,
                         verbose = verbose)
                    if(is.null(gm)){
@@ -104,12 +113,30 @@ r_repos_data <- function(include=NULL,
                    }
                 }, error=function(e){messager(e,v=verbose);NULL})
             }) |> data.table::rbindlist(fill = TRUE)
-            by <- base::intersect(names(meta_gh),names(pkgs))
+            #### Merge back with main data ####
+            ## Need to do this carefully bc there's MANY columns in meta_gh,
+            ## and some accidentalty merging a same-named columns can result 
+            ## in rows with NAs.
+            meta_gh <- add_owner_repo(dt = meta_gh)
+            overlap <- intersect(names(meta_gh),names(pkgs))
+            by <- "ref"
+            overlap <- overlap[overlap!=by]
             pkgs <- data.table::merge.data.table(pkgs,
-                                                 meta_gh,
+                                                 meta_gh[,-overlap, with=FALSE],
                                                  all = TRUE,
                                                  by = by)
         } 
+    } 
+    #### Add hex ####
+    if(isTRUE(add_hex)){
+        messager("Finding hex sticker URLs.",v=verbose)
+        pkgs[,hex_url:= rworkflows::get_hex(refs = ref, 
+                                            paths = NULL,
+                                            add_html = FALSE,
+                                            hex_path = hex_path,
+                                            verbose = TRUE)]
+        pkgs[,hex_url:=sapply(hex_url, function(h){if(is.null(h))NA else h})]
     }
+    #### Return ####
     return(pkgs)
 }